+++ CHANGELOG.md
@@ -1,0 +1,4 @@
+# [2.0.0] - 2024-08-04
+
+## Breaking Changes
+- **Peer Dependencies**: Dropped support for React versions 0.14.0 and 15.x. The library now requires React versions ^16.14.0, ^17.0.0, or ^18.0.0. Projects using this library must update to these versions to avoid compatibility issues.
