+++ examples/src/Countdown.scss
@@ -1,0 +1,55 @@
+.countdown {
+  display: flex;
+  justify-content: center;
+
+  .indicator-countdown {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    text-align: center;
+    position: absolute;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    margin: 0 auto;
+    font-size: 1.3em;
+  }
+
+  .caption {
+    font-size: 0.8em;
+    font-weight: 100;
+    color: #777;
+    margin-bottom: 5px;
+    transition: 0.3s ease;
+
+    &.big {
+      margin-bottom: 0;
+      font-size: 1.1em;
+      color: indianred;
+      animation: blinking 1s step-start 0s infinite;
+
+      span {
+        display: none;
+      }
+    }
+  }
+
+  .time {
+    font-size: 1.6em;
+    font-weight: 100;
+    max-height: 100px;
+    transition: 0.3s ease;
+
+    &.hidden {
+      max-height: 0;
+      overflow: hidden;
+      color: white;
+    }
+  }
+
+  @keyframes blinking {
+    50% {
+      opacity: 0;
+    }
+  }
+}
