+++ examples/src/CounterClockwise.tsx
@@ -1,0 +1,35 @@
+import React from 'react';
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-expect-error
+import ProgressBar from 'react-customizable-progressbar';
+import { ExampleProps } from './Examples';
+
+const CounterClockwise = ({ progress }: ExampleProps) => (
+  <div className="item">
+    <div className="title">
+      <span>Counter-clockwise</span>
+      <a
+        href="https://github.com/martyan/react-customizable-progressbar/blob/master/examples/CounterClockwise.tsx"
+        target="_blank"
+      >
+        code <i className="fa fa-external-link-square" />
+      </a>
+    </div>
+
+    <ProgressBar
+      radius={100}
+      progress={progress}
+      strokeWidth={18}
+      strokeColor="#a0d468"
+      strokeLinecap="round"
+      trackStrokeWidth={18}
+      counterClockwise
+    >
+      <div className="indicator">
+        <div>{progress}%</div>
+      </div>
+    </ProgressBar>
+  </div>
+);
+
+export default CounterClockwise;
