+++ examples/src/Examples.scss
@@ -1,0 +1,114 @@
+@import url('https://fonts.googleapis.com/css?family=Roboto:300,400,500');
+@import url('https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css');
+
+.examples {
+  font-family: 'Roboto', sans-serif;
+
+  .slider {
+    max-width: 280px;
+    display: none;
+    margin: 30px auto;
+    justify-content: space-between;
+    align-items: center;
+
+    .desc {
+      display: inline-block;
+      font-size: 1.05em;
+      color: #777;
+      font-weight: 300;
+    }
+
+    input {
+      display: inline-block;
+    }
+
+    @media (min-width: 480px) {
+      display: flex;
+    }
+  }
+
+  .list {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+
+    .item {
+      flex-basis: 236px;
+      flex-shrink: 0;
+      flex-grow: 0;
+      padding: 5px 15px;
+      background-color: white;
+      border-radius: 5px;
+      margin: 30px;
+
+      .title {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        padding: 15px 5px;
+        margin-bottom: 20px;
+        font-weight: 400;
+        text-transform: uppercase;
+        color: #666;
+        border-bottom: 1px solid #f5f5f5;
+        font-size: 0.9em;
+
+        span {
+          flex: 1;
+          margin-right: 10px;
+        }
+
+        a {
+          text-decoration: none;
+          color: #ccc;
+          font-size: 0.9em;
+        }
+      }
+
+      .indicator {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        text-align: center;
+        position: absolute;
+        top: 0;
+        width: 100%;
+        height: 100%;
+        margin: 0 auto;
+        font-size: 2.2em;
+        font-weight: 100;
+        color: #555;
+        user-select: none;
+      }
+
+      .indicator-volume {
+        display: flex;
+        align-items: flex-end;
+        justify-content: center;
+        text-align: center;
+        position: absolute;
+        top: 0;
+        width: 100%;
+        height: 100%;
+        margin: 0 auto;
+        font-size: 1.3em;
+
+        .inner {
+          margin-bottom: 30px;
+        }
+
+        .percentage {
+          font-size: 1.6em;
+          color: #bbb;
+          font-weight: 100;
+        }
+
+        .icon {
+          font-size: 3em;
+          color: #5d9cec;
+          margin-bottom: 15px;
+        }
+      }
+    }
+  }
+}
