+++ package.json
@@ -1,73 +1,57 @@
 {
-    "name": "react-customizable-progressbar",
-    "version": "1.2.0",
-    "main": "./dist/ReactCustomizableProgressbar.js",
-    "types": "./dist/ReactCustomizableProgressbar.d.ts",
-    "files": [
-        "dist"
-    ],
-    "scripts": {
-        "dev": "webpack serve --config webpack.config.dev.js",
-        "prod": "node ./node_modules/webpack/bin/webpack.js --progress --config ./webpack.config.prod.js"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/martyan/react-customizable-progressbar.git"
-    },
-    "description": "Customizable circular SVG progress bar component for React",
-    "keywords": [
-        "react",
-        "circular",
-        "progress",
-        "bar",
-        "component",
-        "customizable",
-        "svg",
-        "animation"
-    ],
-    "homepage": "https://martinjuzl.com/react-customizable-progressbar",
-    "author": "Martin Juzl",
-    "license": "MIT",
-    "devDependencies": {
-        "@babel/core": "^7.12.3",
-        "@babel/plugin-proposal-class-properties": "^7.12.1",
-        "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
-        "@babel/polyfill": "^7.12.1",
-        "@babel/preset-env": "^7.12.1",
-        "@babel/preset-react": "^7.12.1",
-        "@types/react": "^16.9.55",
-        "@types/react-dom": "^16.9.9",
-        "@types/react-hot-loader": "^3.0.0",
-        "@types/webpack-env": "^1.15.3",
-        "autoprefixer": "^9.8.6",
-        "babel-eslint": "^9.0.0",
-        "babel-loader": "^8.1.0",
-        "babel-plugin-react-css-modules": "^5.2.6",
-        "clean-webpack-plugin": "^0.1.19",
-        "css-loader": "^6.2.0",
-        "eslint": "^7.31.0",
-        "eslint-loader": "^4.0.2",
-        "eslint-plugin-react": "^7.21.5",
-        "html-webpack-plugin": "^5.3.2",
-        "moment": "^2.29.1",
-        "node-sass": "^7.0.0",
-        "postcss-loader": "^6.1.1",
-        "postcss-scss": "^2.1.1",
-        "prop-types": "^15.7.2",
-        "react": "^16.14.0",
-        "react-dom": "^16.14.0",
-        "react-hot-loader": "^3.0.0-beta.7",
-        "sass-loader": "^13.0.2",
-        "style-loader": "^3.2.1",
-        "terser-webpack-plugin": "^5.1.4",
-        "ts-loader": "^9.2.3",
-        "typescript": "^4.0.5",
-        "webpack": "^5.46.0",
-        "webpack-cli": "^4.7.2",
-        "webpack-dev-middleware": "^5.0.0",
-        "webpack-dev-server": "^4.0.0"
-    },
-    "peerDependencies": {
-        "react": "^0.14.0 || ^15.0.0 || ^16.0.0 || ^17.0.0 || ^18.0.0"
-    }
+  "name": "react-customizable-progressbar",
+  "version": "2.0.0",
+  "description": "Customizable circular SVG progress bar component for React",
+  "main": "./dist/index.cjs.js",
+  "module": "./dist/index.esm.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "lint": "eslint src --ext ts,tsx",
+    "build:library": "vite build --config vite.config.library.ts",
+    "dev:examples": "vite --config vite.config.examples.ts"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/martyan/react-customizable-progressbar.git"
+  },
+  "keywords": [
+    "react",
+    "circular",
+    "progress",
+    "bar",
+    "component",
+    "customizable",
+    "svg",
+    "animation"
+  ],
+  "homepage": "https://martinjuzl.com/react-customizable-progressbar",
+  "author": "Martin Juzl",
+  "license": "MIT",
+  "devDependencies": {
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "@typescript-eslint/eslint-plugin": "^8.0.0",
+    "@typescript-eslint/parser": "^8.0.0",
+    "@vitejs/plugin-react": "^3.1.0",
+    "autoprefixer": "^10.4.12",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react": "^7.32.1",
+    "moment": "^2.30.1",
+    "postcss": "^8.4.16",
+    "postcss-loader": "^7.0.1",
+    "prettier": "^3.3.3",
+    "sass": "^1.77.8",
+    "typescript": "^5.2.2",
+    "vite": "^4.5.0",
+    "vite-plugin-eslint": "^1.6.0"
+  },
+  "peerDependencies": {
+    "react": "^16.14.0 || ^17.0.0 || ^18.0.0",
+    "react-dom": "^16.14.0 || ^17.0.0 || ^18.0.0"
+  }
 }
