+++ postcss.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-    plugins: [
-        require('autoprefixer')
-    ]
-}
+  plugins: {
+    autoprefixer: {},
+  },
+};
