+++ tsconfig.json
@@ -1,22 +1,19 @@
 {
-    "compilerOptions": {
-        "declaration": true,
-        "declarationDir": "dist",
-        "outDir": "./dist",
-        "target": "es2015",
-        "module": "es2015",
-        "moduleResolution": "node",
-        "sourceMap": true,
-        "lib": [
-            "dom",
-            "es2015"
-        ],
-        "jsx": "react",
-        "strict": true,
-        "noImplicitAny": true,
-        "esModuleInterop": true,
-        "skipLibCheck": true
-    },
-    "include": ["src"],
-    "exclude": ["node_modules", "dist"]
+  "compilerOptions": {
+    "target": "ESNext",
+    "module": "ESNext",
+    "jsx": "react-jsx",
+    "strict": true,
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true,
+    "outDir": "dist",
+    "declaration": true,
+    "declarationMap": true,
+    "sourceMap": true,
+    "moduleResolution": "node",
+    "resolveJsonModule": true
+  },
+  "include": ["src", "examples"],
+  "exclude": ["node_modules", "dist"]
 }
